Fix package discovery in setup.py (#209)
Added listing all packages for a complete installation using pip (see https://stackoverflow.com/questions/15368054/import-error-on-installed-package-using-setup-py/15368107#15368107 for more details)

Without this the package is unusable - `hloc.utils` and other packages were not being copied to site-packages. I suspect the example notebooks work because the directory is changed to the repo root.